backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
numpy
Pillow
pdfplumber
paddleocr
